Add local image update
Adds a makefile target for development that instead of pushing to
docker hub, builds the image, saves it, then copies it to all workers,
loads it there, and restarts the operator pod.

We copy it to all the workers in case the pod is resurrected on
another node.

Changes the imagepullpolicy to ensure that the local image is used.

Depending on the number of workers, this might be faster than pushing
to docker hub.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>