Support for relating caasapplications as a single unit
Enables relation watching and relation hook firing for caasoperators.

This includes a hack to work around the fact that caasoperator is
built on the Uniter which assumes a 1:1 correspondence with units.
We simply hardcode the unit 0 tag in a few places. Caasoperator will
only successfully watch changes in unit 0 of a relation.
This should be enough functionality to evaluate charm writing in this
context.

api: adds the 1:1 hack

state: fills in a few places where CAAS support was marked as missing.
       adds Watch() to caasunit
       fixes EnterScope to create the right kind of doc for caas units

worker: re-adds some previously commented out
uniter code for relations, re-adds hook tools for relations.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>